Renamed type files.
